requirements.txt
python-dotenv==0.20.0
websocket-client==1.3.3
